% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OR_95CI.R
\name{OR_95CI}
\alias{OR_95CI}
\title{Wald CI for odds ratio}
\usage{
OR_95CI(coef, se, siglevel, roundto)
}
\arguments{
\item{coef}{coefficient estimate from logistic regression model}

\item{se}{standard error from coefficient estimate of logistic regression model}

\item{siglevel}{significance level (coverage) for coefficient estimate}

\item{roundto}{number of decimals to print out}
}
\value{
OR (CI)
}
\description{
Wald CI for odds ratio
}
\examples{
data(toydata)
model <- glm(y ~ x2, data = toydata, family = binomial)
beta <- summary(model)$coefficients["x2", "Estimate"]
SE.beta <- summary(model)$coefficients["x2", "Std. Error"]
OR_95CI(beta, SE.beta, 0.95, 2)
}
