% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toydata.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{Example data}
\format{
A data frame with 100 rows and 3 variables:
\describe{
  \item{y}{a binary outcome coded as 0 and 1}
  \item{x1}{a continuous predictor varirable}
  \item{x2}{a binary predictor variable}
}
}
\source{
course web-page
}
\usage{
toydata
}
\description{
Dataset to demonstrate how OR function works
}
\references{
Week 7 In-class Assignment.
}
\keyword{datasets}
